refactor: code refactor/review fixes
