+++ src/Shared/Components/SelectPicker/index.ts
@@ -1,2 +1,3 @@
 export { default as SelectPicker } from './SelectPicker.component'
 export * from './type'
+export { getSelectPickerOptionByValue } from './utils'
